Add an option to compile code via Electron child process
Adds a new function `compileElectronCode()` which returns a Promise which resolves in the compiled code buffer. This can be enabled via cli with `-e` or `--electron` or by calling `compileFile()` with the `{electron: true}` option. As a side effect, the `compileFile()` function needs to become asynchronous, now returning a Promise which resolves (as before) with the name of the file that was generated.